Rework reparsing of auto-import modules
Previously we relied on FileChangeManager, which is works pretty much the same as FileSystemWatcher. This was kinda bad, because it meant sneaking behind the back of the VS file editor. For example if you changed a code file with auto-imports enabled and you clicked "No" to dialog that asks you to reload external changes we'd go and reparse the file anyway. This was bad because you could have "mod baz;" in the editor, "mod foo;" on the disk and we'd import foo.rs, ignoring your view of the code.