Pass startup commands to GDB via <Command> sub-elements rather than via the command line.
Mingw-w64's GDB's handling of command line parameters containing spaces is broken.
This can be worked around by escaping quotes, e.g. \"alias gdb=echo\" instead of "alias gdb=echo",
but that in turn causes issues with versions of gdb which are not broken in this regard.
Fortunately, there is another way...